requirements.txt
amqp==2.6.1
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
billiard==3.6.4.0
cached-property==1.5.2
celery==4.4.2
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.11
django-celery-beat==2.1.0
django-cors-headers==3.10.1
django-timezone-field==4.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
importlib-metadata==4.10.0
iniconfig==1.1.1
kombu==4.6.11
mysqlclient==2.1.0
packaging==21.3
Pillow==9.0.0
pluggy==0.13.1
prompt-toolkit==3.0.24
py==1.11.0
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.1
vine==1.3.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.7.0
